+++ Tests/JournalMonitorTests.cs
@@ -208,6 +208,23 @@ namespace Tests
         }
 
         [TestMethod]
+        public void TestJournalPlayerMulticrewChat()
+        {
+            // Test for messages received from multicrew. These are received without a defined key for 'Channel' in the player journal.
+            string line = @"{ ""timestamp"":""2017 - 12 - 06T22: 40:54Z"", ""event"":""ReceiveText"", ""From"":""Nexonoid"", ""Message"":""whats up"" }";
+            List<Event> events = JournalMonitor.ParseJournalEntry(line);
+            Assert.IsTrue(events.Count == 1);
+
+            MessageReceivedEvent event1 = (MessageReceivedEvent)events[0];
+
+            Assert.IsTrue(event1.player);
+            Assert.AreEqual("multicrew", event1.channel);
+            Assert.AreEqual("Crew mate", event1.source);
+            Assert.AreEqual("Nexonoid", event1.from);
+            Assert.AreEqual("whats up", event1.message);
+        }
+
+        [TestMethod]
         public void TestJournalMissionAccepted1()
         {
             string line = @"{ ""timestamp"":""2017-05-05T16:07:37Z"", ""event"":""MissionAccepted"", ""Faction"":""Chick Ek Partnership"", ""Name"":""Mission_Sightseeing_Criminal_BOOM"", ""Commodity"":""$Wine_Name;"", ""Commodity_Localised"":""Wine"", ""Count"":3, ""DestinationSystem"":""HR 7221$MISSIONUTIL_MULTIPLE_FINAL_SEPARATOR;Tupa"", ""Expiry"":""2017-05-06T04:31:24Z"", ""Influence"":""Low"", ""Reputation"":""Med"", ""PassengerCount"":7, ""PassengerVIPs"":true, ""PassengerWanted"":true, ""PassengerType"":""Criminal"", ""MissionID"":134724902 }";